requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
pandas==2.2.1
pyyaml==6.0.1
plotly==5.19.0
requests==2.31.0
supabase==2.3.4 